move duck cog into an extension
