+++ duckbot/cogs/duck/__init__.py
@@ -1,0 +1,5 @@
+from .duck import Duck
+
+
+def setup(bot):
+    bot.add_cog(Duck(bot))
