+++ duckbot/cogs/duck/duck.py
