+++ tests/cogs/duck/test_duck.py
@@ -1,30 +1,25 @@
 import pytest
 from unittest import mock
-from duckbot.cogs import Duck
+from duckbot.cogs.duck import Duck
 
 
 @pytest.mark.asyncio
-@mock.patch("discord.Message")
 @mock.patch("random.random", return_value=0.0001)
-async def test_react_duck_random_passes(message, random):
+async def test_react_duck_random_passes(random, message):
     clazz = Duck(None)
-    msg = await clazz.react_duck(message)
-    assert msg is None
-    message.channel.add_reaction.assert_not_called()
+    await clazz.react_duck(message)
+    message.add_reaction.assert_not_called()
 
 
 @pytest.mark.asyncio
-@mock.patch("discord.Message")
 @mock.patch("random.random", return_value=0.000009)
-async def test_react_duck_random_fails(message, random):
+async def test_react_duck_random_fails(random, message):
     clazz = Duck(None)
-    msg = await clazz.react_duck(message)
-    assert msg is None
+    await clazz.react_duck(message)
     message.add_reaction.assert_called_once_with("\U0001F986")
 
 
 @pytest.mark.asyncio
-@mock.patch("discord.ext.commands.Context")
 async def test_github(context):
     clazz = Duck(None)
     await clazz._Duck__github(context)
@@ -32,7 +27,6 @@ async def test_github(context):
 
 
 @pytest.mark.asyncio
-@mock.patch("discord.ext.commands.Context")
 async def test_wiki(context):
     clazz = Duck(None)
     await clazz._Duck__wiki(context)
